# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
...
        "title": "DocServer API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "0.0.0.0:8080",
    "basePath": "/",
    "paths": {
...
        "/documents": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of documents that the currently logged-in user has access to (either owned or shared with them).\n\nThis endpoint supports powerful filtering, sorting, and pagination using query parameters:\n*   `scope`: Control which documents to see:\n*   `owned`: Only documents you created.\n*   `shared`: Only documents shared with you by others.\n*   `all` (default): Both owned and shared documents.\n*   `content_query`: Filter documents based on their JSON content using a specific query language (details likely in separate documentation or examples). This allows searching within the document data itself. Example: `?content_query=metadata.status eq \"published\"`\n*   `sort_by`: Choose the field to sort results by: `creation_date` (default) or `last_modified_date`.\n*   `order`: Set the sort direction: `asc` (ascending) or `desc` (descending, default).\n*   `page`: For pagination, specify the page number (starts at 1, default is 1).\n*   `limit`: For pagination, specify the number of documents per page (default is 20, max is 100).\n\nExample: `/documents?scope=owned\u0026sort_by=last_modified_date\u0026order=asc\u0026page=1\u0026limit=10` (Get the first 10 oldest modified documents owned by the user).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "List and Search Your Documents",
                "parameters": [
                    {
                        "enum": [
                            "owned",
                            "shared",
                            "all"
                        ],
                        "type": "string",
                        "default": "all",
                        "example": "owned",
                        "description": "Filter by ownership: 'owned', 'shared', or 'all'.",
                        "name": "scope",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "example": "user.name eq \"John Doe\"",
                        "description": "Advanced filter based on document content (specific syntax applies).",
                        "name": "content_query",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "creation_date",
                            "last_modified_date"
                        ],
                        "type": "string",
                        "default": "creation_date",
                        "example": "last_modified_date",
                        "description": "Field to sort results by.",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "example": "asc",
                        "description": "Sorting direction.",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "example": 2,
                        "description": "Page number for pagination (starts at 1).",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "example": 50,
                        "description": "Number of documents per page.",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of documents matching the criteria, along with pagination details (total count, current page, limit).",
                        "schema": {
                            "$ref": "#/definitions/api.GetDocumentsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request: One or more query parameters are invalid (e.g., invalid 'scope', incorrect 'content_query' syntax, non-integer 'page'/'limit').",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: Your access token is missing, invalid, or expired.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error: Something went wrong on the server while retrieving documents.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Allows a logged-in user to create and store a new document.\n\nThe document's `content` can be any valid JSON structure – an object (`{}`), an array (`[]`), a string (`\"\"`), a number, a boolean (`true`/`false`), or `null`.\nThe server automatically assigns a unique ID to the document and records the user who created it (the owner) and the creation/modification timestamps.\nYou must provide your access token for authentication. The request body needs a `content` field containing the JSON data you want to store.\n\nExample Request Body:\n```json\n{\n\"content\": {\n\"title\": \"My First Document\",\n\"body\": \"This is the content.\",\n\"tags\": [\"example\", \"getting started\"]\n}\n}\n```",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "Create a New Document",
                "parameters": [
                    {
                        "description": "The JSON content you want to store in the new document.",
                        "name": "document",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateDocumentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Document Created Successfully. The response body contains the details of the newly created document, including its unique ID.",
                        "schema": {
                            "$ref": "#/definitions/models.Document"
                        }
                    },
                    "400": {
                        "description": "Bad Request: The request body is invalid. It must be valid JSON and contain the required 'content' field.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: Your access token is missing, invalid, or expired. You need to be logged in to create documents.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error: Something went wrong on the server while creating the document (e.g., database error).",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.CreateDocumentRequest": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "content": {
                    "description": "Content can be any valid JSON"
                }
            }
        },
...
        "api.GetDocumentsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Document"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "models.Document": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "Can be any JSON structure or simple text"
                },
                "creation_date": {
                    "description": "UTC",
                    "type": "string"
                },
                "id": {
                    "description": "Unique ID (UUID, dashless)",
                    "type": "string"
                },
                "last_modified_date": {
                    "description": "UTC",
                    "type": "string"
                },
                "owner_id": {
                    "description": "Profile ID of the owner",
                    "type": "string"
                }
            }
        },
...
        "utils.APIError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}